+++ src/encrypt.hpp
@@ -4,7 +4,7 @@
 
 #include "AESmath.hpp"
 
-void encrypt(unsigned char* plainInput, unsigned char* encryptedOutput, unsigned char* key, unsigned int keysize);
+void encrypt(unsigned char* input, unsigned char* output, unsigned char* key, unsigned int keysize);
 void subBytes(unsigned char* state);
 void shiftRows(unsigned char* state);
 void mixColumns(unsigned char* state);
